+++ back/src/routers/userRouter.js
@@ -1,7 +1,7 @@
 import is from "@sindresorhus/is";
 import { Router } from "express";
+import { UserController } from "../contoller/userController";
 import { loginRequired } from "../middlewares/loginRequired";
-import { userAuthService } from "../services/userService";
 import { uploadHandler } from "../utils/multer";
 
 const userAuthRouter = Router();
@@ -92,46 +92,7 @@ const userAuthRouter = Router();
  *            $ref: '#components/schemas/User'
  */
 
-userAuthRouter.post("/user/register", async function (req, res, next) {
-  try {
-    if (is.emptyObject(req.body)) {
-      throw new Error(
-        "headers의 Content-Type을 application/json으로 설정해주세요"
-      );
-    }
-
-    // req (request) 에서 데이터 가져오기
-    const name = req.body.name;
-    const email = req.body.email;
-    const password = req.body.password;
-    const myTopics = req.body.myTopics;
-    const bearName = req.body.bearName;
-    const sex = req.body.sex;
-    const age = req.body.age;
-    const occupation = req.body.occupation;
-
-    // 위 데이터를 유저 db에 추가하기
-    const newUser = await userAuthService.addUser({
-      name,
-      email,
-      password,
-      myTopics,
-      bearName,
-      sex,
-      age,
-      occupation,
-      infoProvider: "User",
-    });
-
-    if (newUser.errorMessage) {
-      throw new Error(newUser.errorMessage);
-    }
-
-    res.status(201).json(newUser);
-  } catch (error) {
-    next(error);
-  }
-});
+userAuthRouter.post("/user/register", UserController.create);
 
 /**
  * @swagger
@@ -164,24 +125,7 @@ userAuthRouter.post("/user/register", async function (req, res, next) {
  */
 
 // 로그인
-userAuthRouter.post("/user/login", async function (req, res, next) {
-  try {
-    // req (request) 에서 데이터 가져오기
-    const email = req.body.email;
-    const password = req.body.password;
-
-    // 위 데이터를 이용하여 유저 db에서 유저 찾기
-    const user = await userAuthService.getUser({ email, password });
-
-    if (user.errorMessage) {
-      throw new Error(user.errorMessage);
-    }
-
-    res.status(200).send(user);
-  } catch (error) {
-    next(error);
-  }
-});
+userAuthRouter.post("/user/login", UserController.login);
 
 /**
  * @swagger
@@ -200,20 +144,7 @@ userAuthRouter.post("/user/login", async function (req, res, next) {
  *              token:
  *                  type: string
  */
-userAuthRouter.post("/user/google-login", async function (req, res, next) {
-  try {
-    const { token } = req.body;
-
-    const user = await userAuthService.socialLoginBy(token);
-    if (user.errorMessage) {
-      throw new Error(user.errorMessage);
-    }
-
-    return res.status(200).send(user);
-  } catch (error) {
-    next(error);
-  }
-});
+userAuthRouter.post("/user/google-login", UserController.googleLogin);
 
 /**
  * @swagger
@@ -236,20 +167,7 @@ userAuthRouter.post("/user/google-login", async function (req, res, next) {
  *                  type: string
  */
 
-userAuthRouter.post("/user/mail", async function (req, res, next) {
-  try {
-    const { email, type } = req.body;
-    const user = await userAuthService.sendMail({ email, type });
-
-    if (user.errorMessage) {
-      throw new Error(user.errorMessage);
-    }
-
-    res.status(200).send(user);
-  } catch (error) {
-    next(error);
-  }
-});
+userAuthRouter.post("/user/mail", UserController.sendMail);
 
 /**
  * @swagger
@@ -269,20 +187,7 @@ userAuthRouter.post("/user/mail", async function (req, res, next) {
  *                  type: string
  */
 
-userAuthRouter.post("/user/check/code", async function (req, res, next) {
-  try {
-    const { code } = req.body;
-    const auth = await userAuthService.checkCode({ code });
-
-    if (auth.errorMessage) {
-      throw new Error(auth.errorMessage);
-    }
-
-    res.status(200).send(true);
-  } catch (error) {
-    next(error);
-  }
-});
+userAuthRouter.post("/user/check/code", UserController.checkCode);
 
 /**
  * @swagger
@@ -306,22 +211,7 @@ userAuthRouter.post("/user/check/code", async function (req, res, next) {
 userAuthRouter.post(
   "/user/update/password",
   loginRequired,
-  async function (req, res, next) {
-    try {
-      const { id, password } = req.body;
-      const updatedUser = await userAuthService.updatePassword({
-        user_id: id,
-        password,
-      });
-      if (updatedUser.errorMessage) {
-        throw new Error(updatedUser.errorMessage);
-      }
-
-      res.status(200).send(updatedUser);
-    } catch (error) {
-      next(error);
-    }
-  }
+  UserController.updatePassword
 );
 
 /**
@@ -336,23 +226,7 @@ userAuthRouter.post(
 userAuthRouter.post(
   "/user/cotton/:exp",
   loginRequired,
-  async function (req, res, next) {
-    try {
-      const user_id = req.currentUserId;
-      const exp = Number(req.params.exp);
-      const updatedUser = await userAuthService.updateExp({
-        user_id,
-        exp,
-      });
-      if (updatedUser.errorMessage) {
-        throw new Error(updatedUser.errorMessage);
-      }
-
-      res.status(200).send(updatedUser);
-    } catch (error) {
-      next(error);
-    }
-  }
+  UserController.updateMyExp
 );
 
 /**
@@ -376,26 +250,7 @@ userAuthRouter.post(
   "/user/img",
   loginRequired,
   uploadHandler.single("img"),
-  async function (req, res, next) {
-    try {
-      if (!req.file) {
-        res.status(400).json({ message: "업로드할 이미지가 없습니다" });
-        return;
-      }
-
-      const user_id = req.currentUserId;
-      const url = req.file.path;
-      const updatedUser = await userAuthService.updateImg({ user_id, url });
-
-      if (updatedUser.errorMessage) {
-        throw new Error(updatedUser.errorMessage);
-      }
-
-      res.status(200).send({ url: url });
-    } catch (error) {
-      next(error);
-    }
-  }
+  UserController.updateMyImg
 );
 
 /**
@@ -418,42 +273,10 @@ userAuthRouter.post(
  */
 
 //userlist 반환
-userAuthRouter.get(
-  "/userlist",
-  loginRequired,
-  async function (req, res, next) {
-    try {
-      // 전체 사용자 목록을 얻음
-      const users = await userAuthService.getUsers();
-      res.status(200).send(users);
-    } catch (error) {
-      next(error);
-    }
-  }
-);
+userAuthRouter.get("/userlist", loginRequired, UserController.readUsers);
 
 //사용자 정보 반환
-userAuthRouter.get(
-  "/user/current",
-  loginRequired,
-  async function (req, res, next) {
-    try {
-      // jwt토큰에서 추출된 사용자 id를 가지고 db에서 사용자 정보를 찾음.
-      const user_id = req.currentUserId;
-      const currentUserInfo = await userAuthService.getUserInfo({
-        user_id,
-      });
-
-      if (currentUserInfo.errorMessage) {
-        throw new Error(currentUserInfo.errorMessage);
-      }
-
-      res.status(200).send(currentUserInfo);
-    } catch (error) {
-      next(error);
-    }
-  }
-);
+userAuthRouter.get("/user/current", loginRequired, UserController.readMy);
 
 /**
  * @swagger
@@ -485,60 +308,7 @@ userAuthRouter.get(
  */
 
 //user 정보 수정
-userAuthRouter.put(
-  "/users/:id",
-  loginRequired,
-  async function (req, res, next) {
-    try {
-      // URI로부터 사용자 id를 추출함.
-      const user_id = req.currentUserId;
-      // body data 로부터 업데이트할 사용자 정보를 추출함.
-      const name = req.body.name ?? null;
-      const password = req.body.password ?? null;
-      const myTopics = req.body.myTopics ?? null;
-      const bearName = req.body.bearName ?? null;
-      const level = req.body.level ?? null;
-      const cotton = req.body.cotton ?? null;
-      const height = req.body.height ?? null;
-      const sex = req.body.sex ?? null;
-      const age = req.body.age ?? null;
-      const occupation = req.body.occupation ?? null;
-      const description = req.body.description ?? null;
-      const exp = req.body.exp ?? null;
-      const alert = req.body.alert || null;
-
-      const toUpdate = {
-        name,
-        password,
-        myTopics,
-        bearName,
-        level,
-        cotton,
-        height,
-        sex,
-        age,
-        occupation,
-        description,
-        exp,
-        alert,
-      };
-
-      // 해당 사용자 아이디로 사용자 정보를 db에서 찾아 업데이트함. 업데이트 요소가 없을 시 생략함
-      const updatedUser = await userAuthService.setUser({
-        user_id,
-        toUpdate,
-      });
-
-      if (updatedUser.errorMessage) {
-        throw new Error(updatedUser.errorMessage);
-      }
-
-      res.status(200).json(updatedUser);
-    } catch (error) {
-      next(error);
-    }
-  }
-);
+userAuthRouter.put("/users/:id", loginRequired, UserController.updateMy);
 
 /**
  * @swagger
@@ -564,24 +334,7 @@ userAuthRouter.put(
  */
 
 //user 정보 반환
-userAuthRouter.get(
-  "/users/:id",
-  loginRequired,
-  async function (req, res, next) {
-    try {
-      const user_id = req.params.id;
-      const currentUserInfo = await userAuthService.getUserInfo({ user_id });
-
-      if (currentUserInfo.errorMessage) {
-        throw new Error(currentUserInfo.errorMessage);
-      }
-
-      res.status(200).send(currentUserInfo);
-    } catch (error) {
-      next(error);
-    }
-  }
-);
+userAuthRouter.get("/users/:id", loginRequired, UserController.readOneById);
 
 /**
  * @swagger
@@ -612,23 +365,7 @@ userAuthRouter.get(
  */
 
 //user 삭제 컴포넌트
-userAuthRouter.delete(
-  "/users/user",
-  loginRequired,
-  async function (req, res, next) {
-    try {
-      const user_id = req.currentUserId;
-      // 전체 정보 삭제
-      await userAuthService.deleteUserAllInfo({ user_id });
-      //유저 삭제하는 메소드 호출
-      await userAuthService.deleteUser({ user_id });
-
-      res.status(200).send();
-    } catch (error) {
-      next(error);
-    }
-  }
-);
+userAuthRouter.delete("/users/user", loginRequired, UserController.deleteMy);
 
 /**
  * @swagger
@@ -664,21 +401,6 @@ userAuthRouter.delete(
  */
 
 // 곰 정보 찾기
-userAuthRouter.get(
-  "/bear/:id",
-  loginRequired,
-  async function (req, res, next) {
-    try {
-      const user_id = req.currentUserId;
-      const bearInfo = await userAuthService.getBearInfo({ user_id });
-      if (bearInfo.errorMessage) {
-        throw new Error(bearInfo.errorMessage);
-      }
-      res.status(200).send(bearInfo);
-    } catch (error) {
-      next(error);
-    }
-  }
-);
+userAuthRouter.get("/bear/:id", loginRequired, UserController.readMyBear);
 
 export { userAuthRouter };
